+++ webhooks/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [typescript v0.3.3] webhooks
 ### Releases
 - [NPM v0.3.3] https://www.npmjs.com/package/@avalanche-sdk/webhooks/v/0.3.3 - webhooks
+
+## 2025-04-30 18:41:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.541.0 (2.595.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.4] webhooks
+### Releases
+- [NPM v0.3.4] https://www.npmjs.com/package/@avalanche-sdk/webhooks/v/0.3.4 - webhooks